requirements.txt
# Telegram Bot
python-telegram-bot==20.7

# Environment variables
python-dotenv==1.0.0

# HTTP requests
aiohttp==3.9.1
requests==2.31.0

# Data handling
pandas==2.1.4
numpy==1.26.2

# API interaction
websockets==12.0
asyncio==3.4.3

# Testing
pytest==7.4.3
pytest-asyncio==0.23.2


# Date/time handling
pytz==2023.3.post1

# For potential ML-based analysis later
scikit-learn==1.3.2  

# For Cron jobs 
APScheduler==3.10.1